Added things, fixed things

+++ src/mem/mem.cpp
@@ -1,15 +1,15 @@
-#include "mem.h"
+#include "mem.hpp"
 
 void* memset(void* dest, int data, size_t size) {
     for (size_t i = 0; i < size; i++)
-        *(int*)(dest + i) = data;
+        *(int*)((size_t)dest + i) = data;
 
     return dest;
 }
 
 void* memcpy(void* source, const void* dest, size_t size) {
     for (size_t i = 0; i < size; i++)
-        *(int*)(source + i) = *(int*)(dest + i);
+        *(int*)((size_t)source + i) = *(int*)((size_t)dest + i);
 
     return source;
 }
